move code into common folder
